Updated feature & fixed existing tag issue
Implements fixes to tags not updating properly upon sending POST request.

This feature is to update user tags on the back-end upon sending a POST request. To integrate this with the existing code, server.ts needs to be merged with the script responsible for actually running the server and we also need to trigger POST requests upon visiting content.