Updated recController with Cosine Similarity
Now uses Cosine Similarity to recommend songs as per request from TA. Tested manually but does not include automated tests. Cosine Similarity does not account for weights; plan to discuss this on next stand-up.

Added documentation. Added console log to results can be seen when running the server. Need to merge this with current implementation of back-end, but having personal issues with the set-up. Will also discuss this in next stand-up.